RAXAJSS-328 BUG FIXED: Decimal Places not considered while calculating BMI

+++ src/registrationextjs4/app/controller/BMI.js
@@ -35,8 +35,8 @@ Ext.define('Registration.controller.BMI', {
 
     //Get values from view registrationbmi and call neccessary methods to calculate bmi & display on screen
     getBMIData: function () {
-        var height_cm = parseInt(Ext.getCmp('heightIDcm').getValue()); //Get height value from view registrationbmi
-        var weight_kg = parseInt(Ext.getCmp('weightIDkg').getValue()); //Get weight value from view registrationbmi
+        var height_cm = parseFloat(Ext.getCmp('heightIDcm').getValue()); //Get height value from view registrationbmi
+        var weight_kg = parseFloat(Ext.getCmp('weightIDkg').getValue()); //Get weight value from view registrationbmi
         var bmiInfo = this.calculateBMI(height_cm, weight_kg); //Calculate bmi
         this.updateBMIDisplay(bmiInfo.status, bmiInfo.bmi); //Update the value displayed
         Ext.getCmp('bmiSliderID').setDisabled(true); //Disable the bmiSlider
